App/backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.9
decorator==5.1.1
dnspython==2.7.0
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
frozenlist==1.5.0
http-ece==1.2.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
packaging==24.2
paho-mqtt==2.1.0
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.0
pure_eval==0.2.3
py-vapid==1.9.2
pycparser==2.22
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
pywebpush==2.0.3
pywin32==308
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.3

pytz==2024.1
